[Del] #568 - DailySoptuneCardCoordinator 제거
